Windows: fix region caching that resulted in bad captures
On Windows, the grabbed region is cached to
prevent useless resources allocations and
speed-up further calls to `.grab()`.

But the check to decide if the cached region
needs an update was comparing sizes only.
And using same region sizes with different
region positions between 2 calls to `.grab()`
were resulting in bad screenshots.