linux: Add failure handling to XOpenDisplay() call
Add an explicit check for the return value of XOpenDisplay().  This
function does not seem to use X11 error handlers, and only returns NULL
when it fails.  Without an explicit check, mss ended up passing this
NULL value to further calls and causing segfaults in libX11.  Now it
triggers an explicit exception instead.