Improve typing
Several places, a POINTER(Display) was declared, when the actual
object was an XID (opaque identifier).  In one place, a
POINTER(XWindowAttributes) was declared, instead of an XID.  This is
also why self._handles.drawable had to be cast.

Also, the XErrorEvent structure had its fields in the wrong order.
This meant that error messages in exception details were incorrect.

I'm keeping self._handles.drawable, even though it's currently the
same as self._handles.root, in anticipation that we may later support
capturing individual windows.