Display more complete X11 error messages
The current error message for errors from X11 only gives the name of
the failing function.  It is often useful to also see the information
contained in the error details.

This is an incomplete implementation - it doesn't handle extensions
well - but it's an improvement.  In the future, we may move to xcffib,
which would mean reworking a lot of this error handling anyway.