Fix the test for missing XRRGetScreenResourcesCurrent
Previously, the test would try to monkey-patch the libXrandr module by
deleting the XRRGetScreenResourcesCurrent function.

However, this actually doesn't work as expected.  It deletes all the
configuration you've assigned to that C function (argument and return
types).  But ctypes.CDLL resolves symbols lazily in __getattr__ via
dlsym.  Deleting the attribute only discards the previously created
_FuncPtr object; the next getattr/hasattr on the CDLL will just call
dlsym again and recreate it.

The existing test's code path seemed to be working, because the
implementation was just testing for AttributeError.  The assumption
had been that if XRRGetScreenResourcesCurrent was deleted, then trying
to access it would raise an AttributeError.  However, what was
actually being raised was when the .contents attribute of the return
value was being accessed: since the restype had been erased, then the
return value was now an int, not a pointer.
XRRGetScreenResourcesCurrent was still getting called, but trying to
read its .contents raised the AttributeError.

The new test now uses a proxy object to genuinely hide
XRRGetScreenResourcesCurrent.

Additionally, the test was split into three.  The tests now will
ensure that the right functions are being called based on whether
XRRGetScreenResourcesCurrent is supported, missing on the client
library, or missing on the server.