More minor typing cleanups
Verify all structures against the headers.

Add comment links to X.Org's header files, and the Xlib Programming
Manual where possible, for all structures.

Note where opaque or unneeded fields are in the C definition but not
reflected in the Python code.

Fix a few other pointers and identifiers to XIDs.

Change some int32s in XWindowAttribute to the platform's int, which is
what's in the headers.

None of these issues affect the correctness of the existing code; this
change is only about clarity.