Improve return value types and error checking
Previously, we would always check to see if the return value was not
equal to the integer 0.  We didn't check for NULL pointers at all: a
ctypes _Pointer holding a NULL value is not equal to the integer 0.
Usually, the X server's error event would cue us into the error
anyway, but this is not always guaranteed.

Some integer return values that we didn't need (or want) to check,
such as XCloseDisplay and XDestroyImage, were declared to return
c_void_p, which is never equal to 0.

The new version has the correct return type declarations, and logic to
check whether a call is considered a success or not based on the type.

This fixes some segfaults in older versions of Python that were
introduced by earlier type fixes.