linux: X11 typing and error message improvements (#418)
* Improve typing

Several places, a POINTER(Display) was declared, when the actual
object was an XID (opaque identifier).  In one place, a
POINTER(XWindowAttributes) was declared, instead of an XID.  This is
also why self._handles.drawable had to be cast.

Also, the XErrorEvent structure had its fields in the wrong order.
This meant that error messages in exception details were incorrect.

I'm keeping self._handles.drawable, even though it's currently the
same as self._handles.root, in anticipation that we may later support
capturing individual windows.

* Decode X11 error messages in the correct locale

XGetErrorText returns strings encoded in the current locale (based on LC_CTYPE), which is not always UTF-8.

* Display more complete X11 error messages

The current error message for errors from X11 only gives the name of
the failing function.  It is often useful to also see the information
contained in the error details.

This is an incomplete implementation - it doesn't handle extensions
well - but it's an improvement.  In the future, we may move to xcffib,
which would mean reworking a lot of this error handling anyway.

* Don't use the same inner and outer quotes in f-strings

This was not allowed prior to Python 3.12.

* Ruff cleanups

* More minor typing cleanups

Verify all structures against the headers.

Add comment links to X.Org's header files, and the Xlib Programming
Manual where possible, for all structures.

Note where opaque or unneeded fields are in the C definition but not
reflected in the Python code.

Fix a few other pointers and identifiers to XIDs.

Change some int32s in XWindowAttribute to the platform's int, which is
what's in the headers.

None of these issues affect the correctness of the existing code; this
change is only about clarity.

* Add a CHANGELOG entry

* Remove test_bad_display_structure

This test is no longer needed, and can be removed.

* Don't assume all errors come with details

Sometimes, an Xlib call might fail but there's nothing in the _ERROR
stack.  This happens if the error was something detected internally by
Xlib, rather than coming from the server.  Make sure that we don't try
to read the details in that case.

* locale.getencoding wasn't added until Python 3.11.

* Improve return value types and error checking

Previously, we would always check to see if the return value was not
equal to the integer 0.  We didn't check for NULL pointers at all: a
ctypes _Pointer holding a NULL value is not equal to the integer 0.
Usually, the X server's error event would cue us into the error
anyway, but this is not always guaranteed.

Some integer return values that we didn't need (or want) to check,
such as XCloseDisplay and XDestroyImage, were declared to return
c_void_p, which is never equal to 0.

The new version has the correct return type declarations, and logic to
check whether a call is considered a success or not based on the type.

This fixes some segfaults in older versions of Python that were
introduced by earlier type fixes.