Improve test coverage
Some tests now take a "backend" parameter.  A fixture will iterate
that through the backends implemented for that OS.

Many tests called mss(display=os.getenv("DISPLAY")).  Rather than add
"backend" parameters to each of them, and maybe need to make other
changes in the future, these now take an "mss_impl" fixture.  This is
a callable that will return the appropriate MSS object.

One reason for this is so that we might add a pyvirtualdisplay or
similar mechanism to these functions, when it's available.  This may
enable some tests to run in conditions they might not otherwise.  It
also may help standardize some of the testing.