XCB: Verify visual, and improve error messages
The MSS object, when it's created, will now ensure that the X11
visual's format is exactly what we think it is: 32bpp, as a 24-bit
BGRx or 32-bit BGRA image, in that order, with no extra scanline
padding, color indexing, etc.  This is true of all modern X servers in
real-world situations, but it's not guaranteed.  For instance,
VNC-based servers can be run in depth 16 with RGB565 formats.

The error messages for protocol-level errors (such as bad coordinates)
are only available if the xcb-util-errors library is installed.
(That's libxcb-errors.so.0, in the libxcb-errors0 package on
Debian/Ubuntu.)  This isn't widely installed, and is mostly useful to
developers, so it's optional.  The connection-level errors (like the
hint to check $DISPLAY) are available everywhere.