linux: add an XCB-based backend (#426)
* Working implementation of XCB-based screenshots

This version works fine, but needs improvements, especially to the
testing.

What has happened:

* A ctypes-based module to use XCB has been added.
* The mss.linux.MSS constructor is now a factory, which takes an
  optional `backend` keyword parameter.
* The previous mss.linux implementation has now been designated as the
  "xlib" backend.  Per #425, this will have "legacy" status, and new
  functionality like XShmGetImage won't be implemented there.
* A new XCB-based backend, named "getimage", has been added.  This is
  functionally the same as the existing "xlib" backend, in almost all
  respects, but has a different implementation.
* The GNU/Linux-specific tests that could easily be made to work with
  the new backend have been adjusted.

Notable improvement: I previously was seeing periodic thread failures
in `test_thread_safety`.  In one scenario (Docker container, Python
3.9, Xvfb), I tested 84% failures in `test_thread_safety`.  With the
XCB-based backend, I saw no failures.  (See also #251 and the test
failure I mentioned in #421.)

The performance is, in my brief testing, the same as the existing xlib
implementation.  This PR is about making clearer code and improving
thread safety, to pave the way for future improvements.

What's left to do (some before merging, some possibly later):

* The way the libraries are initialized needs to be rethought; I'll
  want to talk with @BoboTiG about that.
* Many tests in `test_gnu_linux` used monkey-patches that are specific
  to the Xlib implementation.  I've arranged for those to only use
  the Xlib implementation; I need to add corresponding tests for the
  XCB implementation.
* The tests outside of `test_gnu_linux` need to test both backends, as
  well as the different backends we may add to other OSs.  I need to
  add the right Pytest code to parameterize these.
* This code introduces new automatic memory management aspects that
  need tests assigned to them.
* The bulk of the `xcb.py` code is just hand-translations of the [XCB
  protocol specification XML
  files](https://gitlab.freedesktop.org/xorg/proto/xcbproto).
  Hand-translation can be error-prone (see #418).  Since the XCB specs
  are in XML, then the bulk of this code can be auto-generated, like
  xcffib does.  This can be done manually as needed, before shipping
  the distribution, so end users won't have to do this.  (See #425 for
  a discussion of why we're not using xcffib.)
* Even if we don't auto-generate the code, some of the repetitive
  parts can still be hoisted into higher-order functional programming,
  to reduce the risk of making errors in copying.
* There are some parts that I worsened, to satisfy ruff.  I want to
  revisit those.
* The proof-of-concept code in `xcb.py:main` includes some additional
  functionality should be integrated into the MSS class, either in
  this PR or in separate branches.  We also should delete that code
  before merging.
  * Additional testing on the X11 visual, to make sure it's really in
    the format we expect.  (This is probably the only part we'll merge
    in with this PR.)
  * Support for capturing an individual window, without needing its
    location.
  * Identifying whether the alpha channel is meaningful or not.  (This
    is only relevant with individual-window capture; there aren't many
    screens that can turn transparent!)

Fixes: #425

* Improve test coverage

Some tests now take a "backend" parameter.  A fixture will iterate
that through the backends implemented for that OS.

Many tests called mss(display=os.getenv("DISPLAY")).  Rather than add
"backend" parameters to each of them, and maybe need to make other
changes in the future, these now take an "mss_impl" fixture.  This is
a callable that will return the appropriate MSS object.

One reason for this is so that we might add a pyvirtualdisplay or
similar mechanism to these functions, when it's available.  This may
enable some tests to run in conditions they might not otherwise.  It
also may help standardize some of the testing.

* Ruff cleanups

* Small bug fixes to get the tests back to doing meaningful testing

* XCB: Verify visual, and improve error messages

The MSS object, when it's created, will now ensure that the X11
visual's format is exactly what we think it is: 32bpp, as a 24-bit
BGRx or 32-bit BGRA image, in that order, with no extra scanline
padding, color indexing, etc.  This is true of all modern X servers in
real-world situations, but it's not guaranteed.  For instance,
VNC-based servers can be run in depth 16 with RGB565 formats.

The error messages for protocol-level errors (such as bad coordinates)
are only available if the xcb-util-errors library is installed.
(That's libxcb-errors.so.0, in the libxcb-errors0 package on
Debian/Ubuntu.)  This isn't widely installed, and is mostly useful to
developers, so it's optional.  The connection-level errors (like the
hint to check $DISPLAY) are available everywhere.

* Move the library globals to a smart container class

* Improve testing

Test the cases of missing libraries when using XCB.

Add fixtures to reset the XCB library loader after each testcase.

Add fixtures to check for Xlib errors after each testcase.

* Use generated code to interface with XCB instead of hand-written code

* Fix bug in multithreaded initialization guard

* Improve testing and comments

* Add src/xcbproto/README.md

* Fix distribution list to include README.md