Ensure that an X11 connection is open during the test session.
Under X11, when the last client disconnects, the server resets.  If
a new client tries to connect before the reset is complete, it may fail.
Since we often run the tests under Xvfb, they're frequently the only
clients.  Since our tests run in rapid succession, this combination
can lead to intermittent failures.

To avoid this, we open a connection at the start of the test session
and keep it open until the end.