feat(linux): Implement an XShmGetImage-based backend (#431)
* Add XCB MIT-SHM support, and factor out the XCB setup

This only adds the support for the XCB MIT-SHM extension to mss's
internal xcb libraries. The actual usage of shared memory for
screenshots will be done in a future commit.

* Implement an XShmGetImage-based backend.

This is close to complete, but there's a few things that need to be
chased down: notably, the test_thread_safety test is failing, for some
reason.  It also currently doesn't work correctly if the root window
size increases.

That said, this is quite promising: on my computer, the new backend
can take 4k screenshots at 30-34 fps, while the XGetImage backend
could only run at 11-14 fps.

* Expand the xcb-util-errors test to include xshmgetimage

* Track down the closed memfd.

Previously, I had been trying to close the memfd in _shutdown_shm, and
ignoring EBADF.  It turns out that XCB will close the memfd when you
send it to the X server.

I think this was one potential cause of the issues I saw in
test_thread_safety: the two threads would be reallocated each others'
fds, leading to thread A closing an fd that thread B was using,
thinking that it was thread A's memfd.

Fix so that the memfd is only explicitly closed in an error situation.

* Ensure that an X11 connection is open during the test session.

Under X11, when the last client disconnects, the server resets.  If
a new client tries to connect before the reset is complete, it may fail.
Since we often run the tests under Xvfb, they're frequently the only
clients.  Since our tests run in rapid succession, this combination
can lead to intermittent failures.

To avoid this, we open a connection at the start of the test session
and keep it open until the end.

* Update the docs

Also, before marking XShmGetImage as unavailable if the first grab()
fails, also test to see if the fallback XGetImage also fails (such as
if the user gave an out-of-bounds rect).  In that case, just reraise
the exception and try XShmGetImage again with the next grab().

* Changes that should have been in the previous commit

* Change the default backend on Linux to xshmgetimage

Also, fix a minor error in the docs formatting.

* Add a --backend argument to the CLI