Fix bugs in previous autodoc commits.
First, we previously patched ctypes.WINFUNCTYPE in any time
mss.windows was imported, so that autodoc could import mss.windows.
Instead, move that to just when we're running Sphinx.

Second, when I added _close_impl, I accidentally deleted the super()
call in xshmgetimage.  This mean that it never closes connections, and
exhausts the server's client limit.

Finally, in a previous commit, I tested changing mss.lock from a Lock
to an RLock while I was dealing with the fact that MSSBase.close holds
the lock, and the xlib implementation grabs the global lock.  That was
not a great solution, so I backed that out.  Instead, the xlib
implementation just trusts that it's run under the lock.

However, a better implementation might be to create a per-object lock
in MSSBase to protect _closing, instead of using the course-grained
global lock.