feat: add demos (#444)
* Initial commit: TinyTV streaming demo

This is a demo that will probably appeal to few users, but working on
it has helped me with some of the pipeline concepts I'm using in
another demo I'm working on: one to save to an AVI file.

The demo itself seems ready to go, but I still need to add references
to the demos in the documentation.

* Use time.monotonic() instead of time.clock_gettime()

This makes this compatible with Windows and macOS, where
time.clock_gettime() does not exist.