fix: check for negative coordinates (#445)
* Add a test for tuples that have negative coordinates.

Specifically, a negative width or height.

Particularly, this seems like an easy case for a user to hit if
somebody forgets that the PIL tuple style uses top, left, RIGHT,
BOTTOM and not top, left, WIDTH, HEIGHT.

Ask how I know.

Test that we do handle that gracefully.

* Ensure that grab regions are valid

* Remove a now-obsolete test

I thought I'd seen a test for for invalid regions somewhere before...